requirements.txt
PyAudio>=0.2.13
webrtcvad-wheels>=2.0.11
openai-whisper>=20231117
anthropic>=0.50.0
piper-tts>=1.2.0
openwakeword>=0.6.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
